Refactoring through ESLint, second pass.
