+++ src/HorizontalLines/HorizontalLineRegular/index.jsx
@@ -2,6 +2,8 @@ import React from 'react';
 import classnames from 'classnames';
 import styles from './regular-horizontal-line.scss';
 
-const HorizontalLineRegular = () => <span className={classnames(styles.hr)} />;
+function HorizontalLineRegular() {
+  return <span className={classnames(styles.hr)} />;
+}
 
 export default HorizontalLineRegular;
