+++ src/HorizontalLines/HorizontalLineSeparator/index.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import classnames from 'classnames';
 import styles from './horizontal-line-separator.scss';
 
-const HorizontalLineSeparator = () => (
-  <span className={classnames(styles['hr-separator'])} />
-);
+function HorizontalLineSeparator() {
+  return <span className={classnames(styles['hr-separator'])} />;
+}
 
 export default HorizontalLineSeparator;
