+++ src/Icon/IconCustomDot/index.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import classnames from 'classnames';
 import styles from './icon-custom-dot.scss';
 
-const IconCustomDot = () => (
-  <span className={classnames(styles['icon-custom-dot'])} />
-);
+function IconCustomDot() {
+  return <span className={classnames(styles['icon-custom-dot'])} />;
+}
 
 export default IconCustomDot;
