+++ src/Icon/IconHeader/index.jsx
@@ -1,3 +1,7 @@
+/* eslint-disable react/prop-types */
+/* eslint-disable jsx-a11y/no-static-element-interactions */
+/* eslint-disable jsx-a11y/click-events-have-key-events */
+
 import React from 'react';
 import classnames from 'classnames';
 import styles from './icon-header.scss';
