+++ src/Icon/IconLegend/index.jsx
@@ -1,9 +1,13 @@
+/* eslint-disable react/prop-types */
+/* eslint-disable jsx-a11y/no-static-element-interactions */
+/* eslint-disable jsx-a11y/click-events-have-key-events */
+
 import React from 'react';
 import classnames from 'classnames';
 import styles from './icon-legend.scss';
 import IconAction from '../IconAction';
 
-const IconLegend = ({ iconColor, buttonIconType, title, legendType }) => {
+function IconLegend({ iconColor, buttonIconType, title, legendType }) {
   const cn = classnames(styles['icon-legend'], styles[legendType || '']);
   return (
     <span className={cn}>
@@ -17,6 +21,6 @@ const IconLegend = ({ iconColor, buttonIconType, title, legendType }) => {
       )}
     </span>
   );
-};
+}
 
 export default IconLegend;
